chore(caption): create component and service folders
import functions to main js so only one js file is connected to html
