+++ client/caption/components/caption.html
@@ -7,10 +7,10 @@
         <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 
-    <link rel="stylesheet" href="../shared/global.css">
-    <link rel="stylesheet" href="../caption/caption.css">
+    <link rel="stylesheet" href="../../shared/global.css">
+    <link rel="stylesheet" href="../components/caption.css">
     <!-- CSS Stylesheet -->
-    <script src="./caption.js" type="module"></script>
+    <script src="../main.caption.js" type="module"></script>
 
     <!-- Google Fonts Links -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
