+++ client/index.html
@@ -45,7 +45,7 @@
                       <a class="nav-link active" aria-current="page" href="index.html">Home</a>
                   </li>
                   <li class="nav-item">
-                      <a class="nav-link" href="./caption/caption.html">Captions</a>
+                      <a class="nav-link" href="../client/caption/components/caption.html">Captions</a>
                   </li>
               
               </ul>
